Made it so the buttons show in order: Open Channel -> Start Run.
Consolidated the leaderEmbed interactions into the RSA interaction 
handler. The leaderEmbed (shown in the channel where ;afk is used) also 
displays the buttons in the Open Channel -> Start Run order.

The afks display "To join, click here" but it's a well known issue that 
you can't actually join by clicking here. This happens when a raider 
clicks the channel link in a message while it's locked. It binds for 
some reason and the raider has to join ANY other channel on the sidebar 
before they can join through the message link. Therefore, I just removed 
any mentions of the channel in the messages until the raider perms have 
been added to it.