Rework for how active-channel interactions work. It uses big sexy and too the point buttons. This should reduce confusion espcially amongst the ARLs.
With the rework RLs can no longer lock each other out of deleting 
channels with their reacts.

If a channel is deleted while an AFK is up then the run is aborted. The 
person who closed the channel and caused the run to be aborted is 
publically shamed.

Aborted runs have their reacts cleaned so that RSA is cleaner.

I swapped the buttons on the leader embed around so the green button is 
first.