comparePositionTo returns (-inf, inf) so if the role being compared is lower its a negative number, if its the same its 0, if its higher then its positive. Right now the if statement for comparePositionTo checks if the return is 0 or not. Therefore, LOL is considered to be the only non-staff member. Anything less than or greater than LOL is considered to be staff. I added a check for comparePositionTo >= 0 to fix that.
As well, the warn would only proc if the person warning has NO upper 
staff roles. If the warner has any upper staff roles then .size on the 
filter returns a non-zero number and triggers the if. Added a ! in front 
of the upper staff role count check to fix this.