Modded Pops button to afk check
This will add a new button "Modded" and allows a staff member to press it to make all of the keys into a modded key. Automatically logging points and moddedPops for the staff member.

Added a new extension for discord.Message that allows a user to do message.confirmButton and it automatically does confirmation for you.

Added a modded key role to the testing server.

Inside setup.js just formatted some of the lists at the top better