More helpful ;restart, fix potential memory leak
## restart.js
- Added an embed stating which AFK checks & in which servers the bot is waiting for to restart. This also resolves an issue where all afkChecks were thought of as inactive because `afkChecks[].active` is (almost) always undefined.

## afkTemplate.js
- Replaced references of `bot.afkChecks[].active` with `bot.afkModules[].active`.
- Added deletes to `bot.afkModules` for raid deletion commands

I want to note that during testing, and while I can't remember now if it was due to a change I made or something prior, on like 536 `bot.afkChecks[raidID].guild` was null causing afkTemplate to throw. I was originally going to throw a null chain to resolve it but if guild is null there's probably an underlying reason and I'd like to wait to see an error log from prod to both confirm it's an issue and to figure out the cause.

## afkCheck.js
- Sometimes an afkCheck wouldn't show up in afkModules, so made sure in `start` and `loadBotAfkCheck` to add the afkCheck to `bot.afkModules`.
- Added a getter for `afkCheck.guild` so that it can be accessible through afkModules for Restart's embed.
- Added slight dev logging to show how many afk checks are restored after restart.
- Added a delete to `afkModules` when an afkCheck is supposed to be deleted. This is possibly one of the culprits to ViBot's memory leaks as afkModules would perpetually grow and hold references to old afkChecks, forcing them to not be GC'd.