Update utils.js
`LegacyCommandOptions::getMember` was returning a Promise for using Discord id as parameter. Changed to using `Message.Guild.prototype.findMember` in `extensions.js`. This also removes the depedency of message.mentions which is very bad, if they use ign argument and a mention in text it used the mention as reason.