Vote Rework
**Vote Rework**
Version 8.10.0

Instead of submitting a list of users for which their roles will be automatically calculated and a vote will be put up, we switched from that, since we no longer use a binary role progression system, we are much more diverse utilizing around 19 banner roles.

Due to the amount of complexity in the banner system, the vote command needs a rework.

The syntax is
;vote <role> [<user1> (user2) (user3)...]

You input a role, alongside a list of leaders to have their vote be put up for said role.
The channel will be automatically set to the channel you type the command in.

The amount of feedbacks shown is automatically capped at 5.

When you issue the command, you'll be given a "vote configuration" embed, which will allow you to change the channel for the votes, role for which the vote gets put up for, and the cap for the amount of feedbacks shown (1-10)

The embeds are drastically better, with much better designs, and finally introducing images!